examples/jet-stream/message-sender/requirements.txt
aiomonitor
nats-py